requirements.txt
pyg_lib 
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv

--find-links https://data.pyg.org/whl/torch-2.0.0+cu117.html
